refactor: 70 - move function from ui-config
