+++ src/main.ts
@@ -1,1 +1,4 @@
 import './style.css';
+import { RegistrationPage } from './pages/registration';
+
+RegistrationPage();
